requirements.txt
requests
beautifulsoup4
argparse
sublist3r
python-nmap